# decoder: improve default format