--- a/decoder/src/log/mod.rs
+++ b/decoder/src/log/mod.rs
@@ -1,4 +1,6 @@
-//! This module provides interoperability utilities between [`defmt`] and the [`log`] crate.
+//! Handles the conversion of defmt frames to log messages (as strings).
+//! 
+//! Also provides interoperability utilities between [`defmt`] and the [`log`] crate.
 //!
 //! If you are implementing a custom defmt decoding tool, this module can make it easier to
 //! integrate it with logs produced with the [`log`] crate.