@andresovela 2024-06-06
Added `FormatterFormat::Legacy` as an option to retain the old format.