@jonathanpallant 2024-06-10
They seem happy for it to go into 0.3.12, as they are pinned to `==0.3.11` anyway. @Dirbaio said:

> I don't think i'd expect the cli output for logging to be stable? so 0.3.x should be okay