@Urhengulas 2024-06-10
I do like the current two line format, but I am honestly fine with any default. @jonathanpallant what is your call?