@jonathanpallant 2024-06-10
After a lot of discussion on the probe.rs matrix chat, we iterated to:

```rust
        const FORMAT: &str                             = "{{[{L}]%bold} {s}}";
        const FORMAT_WITH_LOCATION: &str               = "{{[{L}]%bold} {s}} {({c:bold}:{ff}:{l:1})%dimmed}";
        const FORMAT_WITH_TIMESTAMP: &str              = "{{t:>8} {[{L}]%bold} {s}}";
        const FORMAT_WITH_TIMESTAMP_AND_LOCATION: &str = "{{t:>8} {[{L}]%bold} {s}} {({c:bold}:{ff}:{l:1})%dimmed}";
```

![image](https://github.com/knurling-rs/defmt/assets/18005923/7e1baeee-5122-4354-98ca-ce02812dcf47)