@jonathanpallant 2024-06-10
Ah, there's an issue. `levels:bin/levels.rs:15` isn't clickable in VS Code. Maybe we should change the first `:` to a space.