@andresovela 2024-06-10
I think trying to make the paths clickable is going to be very painful if we don't have the full path there. Sometimes it will work and sometimes it won't. Especially with dependencies from crates.io where the path would be something like:
```
/home/${USER}/.cargo/registry/src/index.crates.io-6f17d22bba15001f/panic-probe-0.3.1/src/lib.rs:104 
```

vscode won't be able to figure out anything just with `panic_probe panic-probe-0.3.1/src/lib.rs:104` I think. In that case I would say showing `panic_probe: src/lib.rs` would be just as good.