@andresovela 2024-06-10
In the end it's just about striking a good balance, but I think having stuff work only sometimes is probably bad UX