@jonathanpallant 2024-06-10
> In the end it's just about striking a good balance, but I think having stuff work only sometimes is probably bad UX

I see your point, but I don't think I want the full path, because they are so long.

For me `src/lib.rs` is no more helpful than `lib.rs` because `lib.rs` is always in `src/`. `{fff}` seemed like a good compromise (`foobar/src/lib.rs`), and makes the links in my test application broadly work - where VS Code can't unambiguously open the file, it gives me a pretty good list.