@jonathanpallant 2024-06-10
> vscode won't be able to figure out anything just with `panic_probe panic-probe-0.3.1/src/lib.rs:104` I think.

Unless you have the panic-probe-0.3.1 folder open, or any file in that folder open. Then it works it out.