@andresovela 2024-06-10
It'd be cool if we somehow could show the full path for files under the binary crate, and only `crate: dir/file.rs` for other files.