@Urhengulas 2024-06-10
Make it clickable was one main reason to always show the full path by default, iirc.

Also `probe-run` did implement path shortening with the `--shorten-path` option (see https://github.com/knurling-rs/probe-run/blob/b4d99b621fc1de71cb2743fb10e015ff2905915b/src/cli.rs#L99-L101 and https://github.com/knurling-rs/probe-run/blob/b4d99b621fc1de71cb2743fb10e015ff2905915b/src/dep/mod.rs).

Maybe we can see if we can port that code to `defmt-decoder` or `probe-rs`?