@andresovela 2024-06-10
It sounds like that'd need to be implemented in probe-rs