@Urhengulas 2024-06-10
Here are the PRs that implemented shorten paths. It includes examples.

- https://github.com/knurling-rs/probe-run/pull/203
- https://github.com/knurling-rs/probe-run/pull/209
- https://github.com/knurling-rs/probe-run/pull/210