@jonathanpallant 2024-06-11
I added a bunch of comments, and pulled out some common code and put it behind a trait.

This now switches the *default* back as it was, so it requires probe-rs to implement some command line flags to a) pick a defmt formatting style, and b) show the user the list of possible defmt formatting styles.